+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/29 14:46:43 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/11/01 15:16:49 by pmarkaid         ###   ########.fr       */
+/*   Updated: 2024/11/04 12:29:05 by pmarkaid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	main(int argc, char **argv)
 	macro = init_macro(macro);
 	if (!macro)
 		return (write(2, "Error\nMalloc failed\n", 20), 1);
-	validate_map_file(argv[1], macro);
+	read_input(argv[1], macro);
 	// eval_file(argv[1]);
 	// map = read_file(argv[1]);
 	// eval_elements(map);
